+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * pool: rework latency tracking ([Yuki Kishimoto])
 * pool: switch from async to sync message sending for `Relay` ([Yuki Kishimoto])
 * sdk: disable all default features ([Yuki Kishimoto])
+* sdk: set `Client::from_builder` as private ([Yuki Kishimoto])
 
 ### Added